Support dx workflow inputs that reference other analyses.  Parse workflow output.  Save metadata.  Use runInputs rather than originalInputs, so that defaults will be taken from the target workflow.

+++ workflow_utils.py
@@ -17,6 +17,7 @@ import os
 import os.path
 import shutil
 import glob
+import collections
 
 import dxpy
 import dxpy.bindings.dxfile_functions
@@ -24,57 +25,88 @@ import dxpy.bindings.dxfile_functions
 import util.cmd
 import util.file
 
-log = logging.getLogger(__name__)
+_log = logging.getLogger(__name__)
 
 def workflow_utils_init():
-    """Install the dependencies"""
+    """Install the dependencies: cromwell and dxpy."""
     subprocess.check_call('conda install cromwell dxpy')
+
+def _pretty_print_json(json_dict):
+    return json.dumps(json_dict, indent=4, separators=(',', ': '))
 
 #
 # given a dx analysis, run it locally
 #
 
-def get_workflow_inputs(workflow_name, t_dir):
+def get_workflow_inputs(workflow_name, t_dir, docker_tag):
     """Run womtool to get the inputs of the wdl workflow"""
-    for f in glob.glob('pipes/WDL/workflows/*.wdl'):
-        shutil.copy(f, t_dir)
-    for f in glob.glob('pipes/WDL/workflows/tasks/*.wdl'):
-        shutil.copy(f, t_dir)
     with util.file.pushd_popd(t_dir):
+        subprocess.check_call('docker run --rm ' + docker_tag + ' tar cf - source/pipes/WDL > wdl.tar', shell=True)
+        subprocess.check_call('tar xvf wdl.tar', shell=True)
+        for f in glob.glob('source/pipes/WDL/workflows/*.wdl'):
+            shutil.copy(f, '.')
+        for f in glob.glob('source/pipes/WDL/workflows/tasks/*.wdl'):
+            shutil.copy(f, '.')
         return json.loads(subprocess.check_output('womtool inputs ' + workflow_name + '.wdl', shell=True))
     
 
-def get_dx_val(val):
-    dx_cache = '/dev/shm/dxcache'
+def _get_dx_val(val, dx_cache = '/dev/shm/dxcache'):
+    """Resolve a dx value: if it is a scalar, just return that;
+    if it is a dx file, fetch the file, cache it locally, and return the path to the file."""
+    
+    print('parsing val: ', val)
     util.file.mkdir_p(dx_cache)
-    if '$dnanexus_link' in val:
-        dxid = val['$dnanexus_link']
-        assert dxid.startswith('file-')
-        descr = dxpy.describe(dxid)
-        cache_file = os.path.join(dx_cache, dxid) + '-' + descr['name']
-        file_size = int(descr['size'])
-        if not os.path.isfile(cache_file) or os.path.getsize(cache_file) != file_size:
-            print('fetching', dxid, 'to', cache_file)
-            if os.path.isfile(cache_file):
-                os.unlink(cache_file)
-            dxpy.bindings.dxfile_functions.download_dxfile(dxid=dxid, filename=cache_file+'.fetching', show_progress=True)
-            if os.path.getsize(cache_file+'.fetching') == file_size:
-                os.rename(cache_file+'.fetching', cache_file)
-            print('fetched', dxid, 'to', cache_file)
+    if isinstance(val, collections.Mapping) and '$dnanexus_link' in val:
+        link = val['$dnanexus_link']
+        if isinstance(link, collections.Mapping):
+            return _get_dx_val(dxpy.DXAnalysis(link['analysis']).describe()['output'][link['stage']+'.'+link['field']])
+        elif link.startswith('file-'):
+            dxid = link
+            descr = dxpy.describe(dxid)
+            cache_file = os.path.join(dx_cache, dxid) + '-' + descr['name']
+            file_size = int(descr['size'])
+            if not os.path.isfile(cache_file) or os.path.getsize(cache_file) != file_size:
+                print('fetching', dxid, 'to', cache_file)
+                # TODO: check that there is enough free space (with some to spare)
+                if os.path.isfile(cache_file):
+                    os.unlink(cache_file)
+                fs_info = os.statvfs(dx_cache)
+                assert file_size < fs_info.f_bsize * fs_info.f_bavail
+                dxpy.bindings.dxfile_functions.download_dxfile(dxid=dxid, filename=cache_file+'.fetching', show_progress=True)
+                if os.path.getsize(cache_file+'.fetching') == file_size:
+                    os.rename(cache_file+'.fetching', cache_file)
+                print('fetched', dxid, 'to', cache_file)
+        else:
+            raise RuntimeError('Unknown dxid {}'.format(dxid))
         return cache_file
     else:
         return val
 
-def run_dx_locally(workflow_name, analysis_dxid, output_dir, docker_tag=None):
-    """Run a dx analysis locally"""
+def _parse_cromwell_output(cromwell_output_str):
+    """Parse cromwell output"""
+    assert cromwell_output_str.count('Final Outputs:') == 1
+    json_beg = cromwell_output_str.index('Final Outputs:') + len('Final Outputs:')
+    json_end = cromwell_output_str.index('\n}\n', json_beg) + 2
+    return json.loads(cromwell_output_str[json_beg:json_end])
+
+def run_dx_locally(workflow_name, analysis_dxid, docker_tag, output_dir):
+    """Run a dx analysis locally, using a specified version of viral-ngs.
+
+    Notes:
+
+       - record the actual git hash (since git tag might change)
+
+    """
+    subprocess.check_call('docker pull ' + docker_tag, shell=True)
+
     analysis = dxpy.DXAnalysis(dxid=analysis_dxid)
     analysis_descr = analysis.describe()
 
     with util.file.tmp_dir('_workflow_copy') as t_dir:
     
-        wdl_wf_inputs = get_workflow_inputs(workflow_name, t_dir)
+        wdl_wf_inputs = get_workflow_inputs(workflow_name, t_dir, docker_tag=docker_tag)
 
-        dx_wf_inputs = { k.split('.')[-1] : v for k, v in analysis_descr['originalInput'].items()}   # check for dups
+        dx_wf_inputs = { k.split('.')[-1] : v for k, v in analysis_descr['runInput'].items()}   # check for dups
         print('\n'.join(dx_wf_inputs.keys()))
         new_wdl_wf_inputs = {}
         for wdl_wf_input, wdl_wf_input_descr in wdl_wf_inputs.items():
@@ -82,14 +114,17 @@ def run_dx_locally(workflow_name, analysis_dxid, output_dir, docker_tag=None):
             wdl_wf_input = wdl_wf_input.split('.')[-1]
             if wdl_wf_input in dx_wf_inputs:
                 print('HAVE', wdl_wf_input, wdl_wf_input_descr, dx_wf_inputs[wdl_wf_input])
+                # TODO: only use that input if 
                 dx_wf_input = dx_wf_inputs[wdl_wf_input]
-                new_wdl_wf_inputs[wdl_wf_input_full] = map(get_dx_val, dx_wf_input) if isinstance(dx_wf_input, list) else get_dx_val(dx_wf_input)
+                new_wdl_wf_inputs[wdl_wf_input_full] = map(_get_dx_val, dx_wf_input) \
+                                                       if isinstance(dx_wf_input, list) \
+                                                          else _get_dx_val(dx_wf_input)
             else:
                 #print('MISSING', wdl_wf_input, wdl_wf_input_descr)
                 assert '(optional' in wdl_wf_input_descr
 
 
-        print(json.dumps(new_wdl_wf_inputs, indent=4, separators=(',', ': ')))
+        print(_pretty_print_json(new_wdl_wf_inputs))
 
         ################# put in the right docker ID!!  and find a place to keep the docker cache.
 
@@ -104,10 +139,13 @@ def run_dx_locally(workflow_name, analysis_dxid, output_dir, docker_tag=None):
                 subprocess.check_call('sed -i -- "s|{}|{}|g" *.wdl'.format('quay.io/broadinstitute/viral-ngs', docker_tag),
                                       shell=True)
 
+            shutil.rmtree(output_dir, ignore_errors=True)
+            assert not os.path.exists(output_dir)
             util.file.mkdir_p(output_dir)
             util.file.mkdir_p(os.path.join(output_dir, 'outputs'))
             util.file.mkdir_p(os.path.join(output_dir, 'logs'))
             util.file.mkdir_p(os.path.join(output_dir, 'call_logs'))
+            #util.file.mkdir_p(os.path.join(output_dir, 'metadata'))
             wf_opts_dict = { "final_workflow_outputs_dir": os.path.join(output_dir, 'outputs'),
                              "final_workflow_log_dir": os.path.join(output_dir, 'logs'),
                              "final_call_logs_dir": os.path.join(output_dir, 'call_logs')
@@ -116,8 +154,17 @@ def run_dx_locally(workflow_name, analysis_dxid, output_dir, docker_tag=None):
             with open('wf_opts.json', 'wt') as wf_opts_file:
                 json.dump(wf_opts_dict, wf_opts_file, indent=4, separators=(',', ': '))
 
-            wdl_result_str = subprocess.check_output('cromwell run ' + workflow_name + '.wdl -i wf.json -o wf_opts.json', shell=True)
-            print('WDL_RESULT_STR=', wdl_result_str)
+            _log.info('Validating workflow')
+            subprocess.check_call('womtool validate -i wf.json ' + workflow_name + '.wdl', shell=True)
+            _log.info('Validated workflow; calling cromwell')
+            wdl_result_str = subprocess.check_output('cromwell run ' + workflow_name + \
+                                                     '.wdl -i wf.json -o wf_opts.json' + \
+                                                     ' -m ' + os.path.join(output_dir, 'metadata'),
+                                                     shell=True)
+            _log.info('Cromwell returned')
+
+            cromwell_output = _parse_cromwell_output(wdl_result_str)
+            #print('WDL_RESULT_STR=', wdl_result_str)
             util.file.dump_file(os.path.join(output_dir, 'wdl_output.txt'), wdl_result_str)
             
                                       
@@ -135,7 +182,9 @@ def full_parser():
 
 
 if __name__ == '__main__':
-    run_dx_locally(workflow_name='assemble_denovo_with_deplete', analysis_dxid='analysis-FJfqjg005Z3Vp5Q68jxzx5q1',
-                   docker_tag='quay.io/broadinstitute/viral-ngs-dev:1.21.2-5-g394c91e-is-1808241104-refine-assembly-add-reporting',
-                   output_dir='/dev/shm/cromwell_out')
+    #print(_pretty_print_json(_parse_cromwell_output(util.file.slurp_file('/dev/shm/cromwell_out/wdl_output.txt'))))
+    if True:
+        run_dx_locally(workflow_name='assemble_denovo', analysis_dxid='analysis-FJfqjg005Z3Vp5Q68jxzx5q1',
+                       docker_tag='quay.io/broadinstitute/viral-ngs',
+                       output_dir='/dev/shm/cromwell_out')
     #util.cmd.main_argparse(__commands__, __doc__)
